# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: Void-WG Control API
  version: 0.1.0
  description: |
    REST API для управления WireGuard VPN: пользователи, серверы (агенты), peer'ы.
    Авторизация — JWT Bearer (HS256). Heartbeat агентов — заголовком `X-Agent-Token`.
servers:
  - url: http://localhost:8080
    description: Local
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    agentToken:
      type: apiKey
      in: header
      name: X-Agent-Token
  schemas:
...
    Peer:
      type: object
      properties:
        id:             { type: string, format: uuid }
        user_id:        { type: string, format: uuid }
        server_id:      { type: string, format: uuid }
        name:           { type: string }
        public_key:     { type: string }
        assigned_ip:    { type: string }
        enabled:        { type: boolean }
        bytes_rx:       { type: integer }
        bytes_tx:       { type: integer }
        last_handshake: { type: string, format: date-time, nullable: true }
        created_at:     { type: string, format: date-time }
    CreatePeerRequest:
      type: object
      required: [server_id, name]
      properties:
        server_id: { type: string, format: uuid }
        name:      { type: string }
    CreatePeerResponse:
      type: object
      properties:
        peer:   { $ref: '#/components/schemas/Peer' }
        config: { type: string, description: "Готовый wg-quick .conf" }
...
paths:
...
  /api/v1/peers:
    get:
      summary: List my peers
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Peer' } }
    post:
      summary: Create peer (provision)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreatePeerRequest' }
      responses:
        '201':
          description: created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreatePeerResponse' }
...
security:
  - bearerAuth: []